move `ReadRecord` and `StreamReading` to `read.rs`
