+++ reqrio/src/stream/mod.rs
@@ -13,24 +13,10 @@ pub use aync::*;
 pub use buffer::{ReadOffset, RecordBuffer};
 pub use proxy::Proxy;
 pub use proxy::ProxyStream;
-use std::io::Write;
-#[cfg(feature = "aync")]
-use std::ops::Range;
+use std::io::{Read, Write};
 use std::path::{Path, PathBuf};
-use std::sync::atomic::{AtomicBool, Ordering};
-use std::sync::Arc;
 use std::{env, io};
-use std::pin::Pin;
-use std::task::{Context, Poll};
-use pin_project_lite::pin_project;
 pub use sync_stream::SyncStream;
-#[cfg(feature = "aync")]
-use tokio::io::{AsyncRead, ReadHalf};
-use tokio::io::ReadBuf;
-use tokio::sync::futures::Notified;
-#[cfg(feature = "aync")]
-use tokio::sync::mpsc::Sender;
-use tokio::sync::Notify;
 pub use ws::{WebSocket, WebSocketBuilder};
 
 
@@ -173,36 +159,31 @@ impl Stream {
         }
     }
 
-    pub fn sync_write(&mut self, buf: &[u8]) -> HlsResult<()> {
-        match self {
-            Stream::SyncHttp(s) => {
-                s.write_all(buf)?;
-                Ok(())
-            }
-            Stream::SyncHttps(s) => {
-                s.write_all(buf)?;
-                Ok(())
-            }
-            _ => Err("Unsupported sync write".into()),
-        }
+    pub fn sync_write(&mut self, buf: &[u8]) -> io::Result<()> {
+        let stream: &mut dyn Write = match self {
+            Stream::SyncHttp(stream) => stream,
+            Stream::SyncHttps(stream) => stream,
+            _ => return Err(io::Error::other("Unsupported sync write")),
+        };
+        stream.write_all(buf)
     }
 
     pub fn sync_read(&mut self, buffer: &mut Buffer) -> HlsResult<()> {
-        match self {
-            Stream::SyncHttp(stream) => {
-                let len = io::Read::read(stream, buffer.unfilled())?;
-                if len == 0 { return Err(HlsError::PeerClosedConnection); }
-                buffer.add_len(len);
-                Ok(())
-            }
-            Stream::SyncHttps(stream) => {
-                let len = io::Read::read(stream, buffer.unfilled())?;
-                if len == 0 { return Err(HlsError::PeerClosedConnection); }
-                buffer.add_len(len);
-                Ok(())
-            }
-            _ => Err("Unsupported async read".into()),
+        let stream: &mut dyn Read = match self {
+            Stream::SyncHttp(stream) => stream,
+            Stream::SyncHttps(stream) => stream,
+            _ => return Err("Unsupported sync write".into()),
+        };
+        if buffer.unfilled_len() == 0 {
+            return Err(BufferError::CapacityTooSmall {
+                needed: 4096,
+                current: buffer.unfilled_len(),
+            }.into());
         }
+        let len = stream.read(buffer.unfilled())?;
+        if len == 0 { return Err(HlsError::PeerClosedConnection); }
+        buffer.add_len(len);
+        Ok(())
     }
 
     pub fn sync_shutdown(&mut self) -> HlsResult<()> {
@@ -211,125 +192,6 @@ impl Stream {
             Stream::SyncHttps(s) => Ok(s.shutdown()?),
             _ => Err("Unsupported async read".into()),
         }
-    }
-}
-
-pin_project! {
-    pub struct ReadRecord<'a, S: Send> {
-        reader: StreamReading<'a, S>,
-        #[pin]
-        notified: Notified<'a>,
-        record_len: usize,
-    }
-}
-
-impl<'a, S: Send> ReadRecord<'a, S> {}
-
-impl<'a, S: AsyncRead + Unpin + Send> Future for ReadRecord<'a, S> {
-    type Output = io::Result<Range<usize>>;
-    fn poll(self: Pin<&mut Self>, cx: &mut Context<'_>) -> Poll<Self::Output> {
-        let mut project = self.project();
-        if *project.record_len == 0 {
-            match project.reader.read_record_size(project.notified.as_mut(), cx)? {
-                Poll::Ready(size) => *project.record_len = size,
-                Poll::Pending => return Poll::Pending,
-            }
-        }
-        if project.reader.buffer.current_size() < *project.record_len &&
-            project.reader.read_size(*project.record_len, project.notified.as_mut(), cx)?.is_pending() {
-            return Poll::Pending;
-        }
-        Poll::Ready(Ok(project.reader.buffer.next_record_offset(*project.record_len)))
-    }
-}
-
-
-struct StreamReading<'a, S: Send> {
-    buffer: &'a mut RecordBuffer,
-    reader: &'a mut ReadHalf<S>,
-}
-
-impl<'a, S: AsyncRead + Unpin + Send> StreamReading<'a, S> {
-    fn read_size(&mut self, want_size: usize, mut notified: Pin<&mut Notified<'_>>, cx: &mut Context<'_>) -> Poll<io::Result<()>> {
-        let current_size = self.buffer.current_size();
-        if current_size >= want_size { return Poll::Ready(Ok(())); }
-        let mut unfilled = loop {
-            match self.buffer.unfilled_mut(want_size, current_size) {
-                Poll::Ready(res) => break ReadBuf::new(res),
-                Poll::Pending => match notified.as_mut().poll(cx) {
-                    Poll::Ready(_) => continue,
-                    Poll::Pending => return Poll::Pending,
-                }
-            }
-        };
-        let need_read_size = want_size - current_size;
-        while unfilled.filled().len() < need_read_size {
-            match Pin::new(&mut self.reader).poll_read(cx, &mut unfilled)? {
-                Poll::Ready(_) => {
-                    if unfilled.filled().is_empty() {
-                        return Poll::Ready(Err(HlsError::PeerClosedConnection.into()));
-                    }
-                }
-                Poll::Pending => {
-                    let size = unfilled.filled().len();
-                    self.buffer.add_filled(size);
-                    return Poll::Pending
-                }
-            };
-        }
-        let size = unfilled.filled().len();
-        self.buffer.add_filled(size);
-        Poll::Ready(Ok(()))
-    }
-
-
-    fn read_record_size(&mut self, notified: Pin<&mut Notified<'_>>, cx: &mut Context<'_>) -> Poll<io::Result<usize>> {
-        if self.buffer.current_size() < 5 && self.read_size(23, notified, cx)?.is_pending() { return Poll::Pending; };
-        Poll::Ready(Ok(self.buffer.next_record_len()))
-    }
-}
-
-
-#[cfg(feature = "aync")]
-pub struct StreamRead<S: Send> {
-    pub buffer: RecordBuffer,
-    pub reader: ReadHalf<S>,
-    pub sender: Sender<io::Result<ReadOffset>>,
-    pub closed: Arc<AtomicBool>,
-    pub notify: Arc<Notify>,
-}
-
-#[cfg(feature = "aync")]
-impl<S: AsyncRead + Unpin + Send> StreamRead<S> {
-    fn read_record(&mut self) -> ReadRecord<'_, S> {
-        ReadRecord {
-            notified: self.notify.notified(),
-            reader: StreamReading {
-                buffer: &mut self.buffer,
-                reader: &mut self.reader,
-            },
-            record_len: 0,
-        }
-    }
-
-
-    pub async fn run(&mut self) {
-        loop {
-            if self.closed.load(Ordering::SeqCst) { break; }
-            match self.read_record().await {
-                Ok(record_offset) => {
-                    // println!("send: {:?}; len: {}", record_offset, record_offset.len() - 5);
-                    let offset = ReadOffset::new(record_offset, &self.buffer, self.notify.clone());
-                    let ret = self.sender.send(Ok(offset)).await;
-                    if ret.is_err() { break; }
-                }
-                Err(e) => {
-                    let _ = self.sender.send(Err(e)).await;
-                    break;
-                }
-            }
-        }
-        println!("stream closed");
     }
 }
